+++ src/containers/gds/GDSDataTable/index.tsx
@@ -11,7 +11,7 @@ import { useToastContext } from '../../../providers/ToastProvider';
 import { ColumnProps } from 'primereact/column';
 import { getStringReadableBytes, showDisplayText } from '../../../utils/util';
 import moment from 'moment';
-import DataActionButton from '../../../components/DataActionButton';
+import DataActionButton from '../../utils/DataActionButton';
 import FilePreviewButton from '../../../components/FilePreviewButton';
 import DataSearchFilterButton from '../../../components/DataSearchFilterButton';
